prediction-service/requirements.txt
fastapi
uvicorn
httpx
pydantic
pika
pydantic-settings
loguru
aiolimiter
python-magic
backoff
redis